refactor(SF-121): split url4.py — AST / grammar / resolve modules
url4.py mixed grammar parsing (TatSu Semantics), AST nodes, async
resolution, HTTP fetch helpers, and backend-call dispatch in one
457-line module. Three focused modules now own each concern; the
original ``url4`` module is a thin re-export facade for backward
compat.

- url4_ast.py (82 lines): Url4Url / Url4RelUrl / Url4Text / Url4List /
  Url4BackendCall / Url4ExpandedSource dataclasses + Url4Node union.
  No httpx / TatSu / plugin-registry deps; consumers that just need
  the AST types pay nothing else.
- url4_grammar.py (190 lines): GRAMMAR PEG, Url4Semantics class,
  parse() — owns the parser + once-compiled cache.
- url4_resolve.py (166 lines): resolve / resolve_str / _fetch_url /
  _fetch_relative / _dispatch_backend_call / _resolve_expanded_source
  — owns network + plugin dispatch.
- url4.py (50 lines): backward-compat shim re-exporting public names.

Internal callers (interpreter, ensemble, resolver) updated to import
from the new modules. Test mock.patch targets updated to point at
url4_resolve._fetch_url / _fetch_relative.

707 plugin unit tests pass.

https://app.asana.com/1/1185126988600652/project/1213628819033917/task/SF-121
